Change throttle from int to float

+++ src/chassis.cpp
@@ -76,7 +76,7 @@ void arcade_drive() {
 }
 
 void curvature_drive() {
-	int power = 
+	float power = 
 	    controller.get_analog(pros::E_CONTROLLER_ANALOG_LEFT_Y) / 127.0;
 	float curvature =
 	    controller.get_analog(pros::E_CONTROLLER_ANALOG_RIGHT_X) / 127.0;
